# repo/backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Hospitality Operations Management System API
  description: Backend API for managing events, resources, reservations, recipes, inventory, entitlements, and operational workflows.
  version: 1.0.0
servers:
  - url: /api
    description: API base path
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from POST /auth/login or POST /auth/refresh.
  schemas:
...
    PaginatedResponse:
      type: object
      properties:
        data:
          type: array
          items: {}
        pagination:
          type: object
          properties:
            cursor:
              type: string
              nullable: true
            hasMore:
              type: boolean
            total:
              type: integer
...
    InventorySnapshot:
      type: object
      properties:
        id:
          type: string
          format: uuid
        snapshot_date:
          type: string
          format: date
        items:
          type: array
          items:
            type: object
            properties:
              item_id:
                type: string
                format: uuid
              expected_quantity:
                type: number
              actual_quantity:
                type: number
              gap:
                type: number
        created_at:
          type: string
          format: date-time
...
paths:
...
  /recipes:
    get:
      tags: [Recipes]
      summary: List recipes
      parameters:
        - name: cursor
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Paginated recipes
    post:
      tags: [Recipes]
      summary: Create a new recipe
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title]
              properties:
                title:
                  type: string
                description:
                  type: string
      responses:
        '201':
          description: Recipe created
  /recipes/{id}:
    get:
      tags: [Recipes]
      summary: Get recipe with versions
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Recipe details with version history
...
  /inventory/snapshots:
    get:
      tags: [Inventory]
      summary: List inventory snapshots
      parameters:
        - name: date
          in: query
          schema:
            type: string
            format: date
        - name: cursor
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Paginated snapshots
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/InventorySnapshot'
...
  /inventory/gaps/{id}/resolve:
    post:
      tags: [Inventory]
      summary: Resolve an inventory gap
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [resolution_method, notes]
              properties:
                resolution_method:
                  type: string
                  enum: [recount, adjustment, write_off]
                notes:
                  type: string
      responses:
        '200':
          description: Gap resolved
...